Add force option to retrive_all_files and retrieve_file methods on ModelDatabase and ModelSnapshot

+++ src/pharmpy/workflows/model_database/local_directory.py
@@ -117,22 +117,27 @@ class LocalDirectoryDatabase(NonTransactionalModelDatabase):
         file_names = [f.name for f in self.path.glob(f'{name}.*') if f.is_file()]
         return file_names
 
-    def retrieve_file(self, name, filename, destination_path):
+    def retrieve_file(self, name, filename, destination_path, force=False):
         # Return path to file
         path = self.path / filename
         if path.is_file() and stat(path).st_size > 0:
+            if (
+                destination_path.is_file() or (destination_path / filename).is_file()
+            ) and force is not True:
+                raise ValueError(f'Cannot retrieve file: {destination_path} already exists')
             shutil.copy2(path, destination_path)
         else:
             raise FileNotFoundError(f"Cannot retrieve {filename} for {name}")
 
-    def retrieve_all_files(self, name, destination_path):
+    def retrieve_all_files(self, name, destination_path, force=False):
         # Retrieve all files stored for one model
         if destination_path.is_file():
             raise ValueError(f"Cannot copy files to `{destination_path}`:  is a file")
 
         file_names = self.list_all_files(name)
         for file_name in file_names:
-            shutil.copy2(self.path / file_name, destination_path)
+            dst = destination_path / file_name
+            self.retrieve_file(file_name, dst, force)
 
     def retrieve_model(self, name):
         filename = name + self.file_extension
@@ -340,25 +345,29 @@ class LocalModelDirectoryDatabaseSnapshot(ModelSnapshot):
         else:
             raise ValueError(f"Cannot find any files for {self.key}")
 
-    def retrieve_file(self, filename, destination_path):
+    def retrieve_file(self, filename, destination_path, force=False):
         # Return path to file
         path = self.database.path / str(self.key) / filename
         if path.is_file() and stat(path).st_size > 0:
+            if (
+                destination_path.is_file() or (destination_path / filename).is_file()
+            ) and force is not True:
+                raise ValueError(f'Cannot retrieve file: {destination_path} already exists')
             shutil.copy2(path, destination_path)
         else:
             raise FileNotFoundError(f"Cannot retrieve {filename} for {self.key}")
 
-    def retrieve_all_files(self, destination_path):
+    def retrieve_all_files(self, destination_path, force=False):
         if not destination_path.exists():
             raise ValueError(
                 f"Cannot copy files to `{destination_path}`:  directory does not exist"
             )
         if destination_path.is_file():
             raise ValueError(f"Cannot copy files to `{destination_path}`:  is a file")
-        path = self.database.path / str(self.key)
         file_names = self.list_all_files()
         for file_name in file_names:
-            shutil.copy2(path / file_name, destination_path)
+            dst = destination_path / file_name
+            self.retrieve_file(file_name, dst, force)
 
     def retrieve_model(self):
         path = self._find_full_model_path()
